+++ Color/src/Graphics/Color/Model/CMYK.hs
@@ -26,7 +26,6 @@ module Graphics.Color.Model.CMYK
   ) where
 
 import Foreign.Storable
-import Graphics.Color.Model.Alpha
 import Graphics.Color.Model.Internal
 import Graphics.Color.Model.RGB