+++ Color/src/Graphics/Color/Model/HSI.hs
@@ -30,7 +30,6 @@ module Graphics.Color.Model.HSI
   ) where
 
 import Foreign.Storable
-import Graphics.Color.Model.Alpha
 import Graphics.Color.Model.Internal
 import Graphics.Color.Model.RGB