+++ Color/src/Graphics/Color/Model/YCbCr.hs
@@ -29,7 +29,6 @@ module Graphics.Color.Model.YCbCr
   ) where
 
 import Foreign.Storable
-import Graphics.Color.Model.Alpha
 import Graphics.Color.Model.Internal
 import Graphics.Color.Model.RGB
 import Graphics.Color.Model.Y