+++ Color/src/Graphics/Color/Space.hs
@@ -14,7 +14,6 @@ module Graphics.Color.Space
   , module X
   ) where
 
-import Graphics.Color.Model.Alpha as X
 import Graphics.Color.Model.Internal as X
 import Graphics.Color.Space.Internal as X
 import Graphics.Color.Space.RGB.Internal as X