+++ Color/src/Graphics/Color/Space/RGB/AdobeRGB.hs
@@ -33,7 +33,6 @@ module Graphics.Color.Space.RGB.AdobeRGB
   ) where
 
 import Foreign.Storable
-import Graphics.Color.Model.Alpha
 import Graphics.Color.Model.Internal
 import qualified Graphics.Color.Model.RGB as CM
 import Graphics.Color.Space