+++ Color/src/Graphics/Color/Space/RGB/Derived/SRGB.hs
@@ -26,7 +26,6 @@ module Graphics.Color.Space.RGB.Derived.SRGB
   ) where
 
 import Foreign.Storable
-import Graphics.Color.Model.Alpha
 import Graphics.Color.Model.Internal
 import qualified Graphics.Color.Model.RGB as CM
 import Graphics.Color.Space